requirements.txt
streamlit>=1.26.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0